requirements.txt
wheel
pandas
torch
numpy
PyMuPDF

pdf2image
Pillow
tqdm
pdf4llm


python-pptx
docx2txt
python-docx 

# textract
# tika